+++ src/plone/app/fhirfield/interfaces.py
@@ -4,24 +4,7 @@ from fhirspec import FHIR_RELEASES
 from plone.app.fhirfield.compat import _
 from plone.schema import JSONField
 from zope import schema as zs
-from zope.interface import Attribute
-from zope.interface import Interface
 from zope.schema.interfaces import IField
-
-
-class IFhirResourceModel(Interface):
-    """ """
-
-    resource_type = Attribute("resource_type", _("Resource Type"))
-    id = Attribute("id", _("Logical id of this artifact."))
-    implicitRules = Attribute(
-        "implicitRules", _("A set of rules under which this content was created.")
-    )
-    language = Attribute("language", _("Language of the resource content."))
-    meta = Attribute("meta", _("Metadata about the resource"))
-
-    def json():
-        """ """
 
 
 class IFhirResource(IField):
@@ -44,21 +27,3 @@ class IFhirResource(IField):
 
     def from_none():
         """Make FhirResourceValue isntance without FHIR data """
-
-
-class IFhirResourceValue(Interface):
-    """ """
-
-    _encoding = Attribute(
-        "_encoding", _("Encoding name that will be used during json generation")
-    )
-    _storage = Attribute("_storage", _("_storage to hold Fhir resource model object."))
-
-    def stringify(prettify=False):
-        """Transformation to JSON string representation"""
-
-    def patch(patch_data):
-        """FHIR Patch implementation: https://www.hl7.org/fhir/fhirpatch.html"""
-
-    def foreground_origin():
-        """Return the original object of FHIR model that is proxied!"""