+++ src/plone/app/fhirfield/tests/__init__.py
@@ -1,21 +1,12 @@
 # _*_ coding: utf-8 _*_
 import os
 
-from plone.app.fhirfield.interfaces import IFhirResourceModel
-
 
 __author__ = "Md Nazrul Islam<email2nazrul@gmail.com>"
 
 
 BASE_TEST_PATH = os.path.dirname(os.path.abspath(__file__))
 FHIR_FIXTURE_PATH = os.path.join(BASE_TEST_PATH, "fixture", "FHIR")
-
-
-class IWrongInterface(IFhirResourceModel):
-    """ """
-
-    def meta():
-        """ """
 
 
 class NoneInterfaceClass(object):